Issue #83 Fix drawing of screen by converting draw_offset/draw_max to short.
There was an issue with `    max_y = screen_state.draw_offset_y + screen_state.draw_max_y;`
where the value was not being calculated correctly. When draw_offset_y was
0, the issue would occur, but setting max_y to draw_max_y worked fine.
There must have been some conversion issue with the two unsigned ints and
stored in an unsigned int.
-e 
Original/build commit ID: 6310f5044e446e593479553f466426510d735c07

+++ screen_state.c
@@ -55,10 +55,10 @@ typedef struct {
     unsigned char *background_tiles;
     unsigned char *background_color_palette;
     unsigned int background_width;
-    unsigned int draw_offset_x;
-    unsigned int draw_offset_y;
-    unsigned int draw_max_x;
-    unsigned int draw_max_y;
+    unsigned short draw_offset_x;
+    unsigned short draw_offset_y;
+    unsigned short draw_max_x;
+    unsigned short draw_max_y;
 } screen_state_t;
 
 #endif
